+++ NEWS.md
@@ -1,3 +1,6 @@
+# tern 0.6.7.9000
+
+* issues with no comments:
 
 # tern 0.6.7